functions/ingestion/requirements.txt
# Ingestion function dependencies
functions-framework==3.9.2
google-cloud-firestore==2.14.0
google-cloud-secret-manager==2.18.0
google-cloud-tasks==2.15.0
google-cloud-storage==2.14.0
